Use CO title instead of Bitstream name when doing comparison & refactoring
Previously, it was incorrectly assumed that DDB file name and CO's bitstream name are the same,
when, instead the CO's title should be used. The CO's Title is now being compared to the
DDB file's title and if that is missing then using the file name; this is being done as this
is the method used when we add an entity in the upsert Lambda.

Test code has been refactored for clarity